# integrations/openapi.yaml
openapi: 3.1.0
info:
  title: conversation-blackboard
  version: 0.3.0
  description: |
    Native HTTP projection for reading, writing, and administering a conversation-blackboard.
    This OpenAPI document is an HTTP projection of Rust application semantics; it is not domain authority.
    Persisted source/instance provenance is resolved by the server; message clients never supply it.
    conversation_ref is optional provider-side provenance on stored messages and carries no authority.
servers:
  - url: https://board.cafefeed.idv.tw
    description: Production conversation-blackboard endpoint.
paths:
...
  /api/access-context:
    get:
      operationId: blackboardAccessContext
      summary: Read the authenticated caller's effective Blackboard access context.
      description: HTTP projection of the shared authorization kernel.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      responses:
        '200':
          description: Effective identity, principal, participant attribution, capabilities, and grants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/executions/{intent_id}:
    get:
      operationId: blackboardExecutionReceipt
      summary: Read one durable semantic execution receipt.
      description: The intent_id is semantic identity. delivery_id remains transport ingress provenance and is intentionally absent from the receipt.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IntentId'
      responses:
        '200':
          description: Durable execution receipt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionReceiptEnvelope'
        '400':
          description: Invalid semantic intent identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No durable receipt exists for this participant and intent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/executions/{intent_id}/audit:
    get:
      operationId: blackboardExecutionAudit
      summary: Read immutable committed execution audit evidence.
      description: Current policy authorizes access to this historical read model; historical authorization contents are read from persisted evidence and are not re-evaluated.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IntentId'
      responses:
        '200':
          description: Immutable historical execution audit bundle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionAuditEnvelope'
        '400':
          description: Invalid semantic intent identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No committed execution exists for this participant and intent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/executions/{intent_id}/audit/integrity:
    get:
      operationId: blackboardExecutionAuditIntegrity
      summary: Verify structural integrity of committed execution audit evidence.
      description: Current policy authorizes access; the report is produced by the canonical read-only integrity verifier and does not re-evaluate historical authorization or repair history.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IntentId'
      responses:
        '200':
          description: Structural integrity report for the requested semantic execution identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionAuditIntegrityEnvelope'
        '400':
          description: Invalid semantic intent identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/execution-audit/sweep:
    get:
      operationId: blackboardExecutionAuditSweep
      summary: Verify database-wide execution audit integrity.
      description: Privileged projection of the canonical read-only execution audit sweep. Invalid integrity state is returned as report data with HTTP 200; the endpoint does not repair history.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      responses:
        '200':
          description: Canonical database-wide execution audit sweep report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionAuditSweepEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    webSessionAuth:
      type: apiKey
      in: header
      name: X-Blackboard-Web-Session
      description: Short-lived page-memory Human Web or Guest session token.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked authentication proof.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated session does not have authority for this resource/action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ParticipantId:
      type: string
      minLength: 1
      maxLength: 64
      description: Durable Blackboard attribution identity. It is not a transport credential.
...
    IntentId:
      type: string
      minLength: 1
      maxLength: 256
      description: Transport-independent semantic operation identity used for idempotency and receipt lookup.
    DeliveryId:
      type: string
      minLength: 1
      maxLength: 256
      description: Transport delivery provenance identity. Distinct deliveries may carry the same semantic intent_id; delivery_id is not part of the semantic execution receipt.
    Principal:
      type: object
      additionalProperties: false
      required: [provider, subject]
      properties:
        provider:
          type: string
        subject:
          type: string
    AccessContext:
      type: object
      additionalProperties: false
      required: [participant_id, principal, capabilities, grants]
      properties:
        identity:
          $ref: '#/components/schemas/Identity'
        principal:
          $ref: '#/components/schemas/Principal'
        participant_id:
          anyOf:
            - $ref: '#/components/schemas/ParticipantId'
            - type: 'null'
        role:
          type: string
        capabilities:
          type: array
          items:
            type: string
        grants:
          type: array
    ExecutionReceipt:
      type: object
      additionalProperties: false
      required: [participant_id, intent_id, intent_hash, capability, message_id, status]
      description: Durable semantic execution receipt. Transport delivery_id is intentionally excluded and remains ingress provenance.
      properties:
        participant_id:
          $ref: '#/components/schemas/ParticipantId'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        intent_hash:
          type: string
        capability:
          type: string
        message_id:
          type: integer
          minimum: 1
        status:
          type: string
    ExecutionReceiptEnvelope:
      type: object
      additionalProperties: false
      required: [execution]
      properties:
        execution:
          $ref: '#/components/schemas/ExecutionReceipt'
    AuthorizationProvenance:
      type: object
      additionalProperties: false
      required: [participant_id, intent_id, principal, capability, resource, source, reason, grant_id]
      description: Historical non-secret authorization decision and admitted request context persisted at execution commit time. Legacy principal/resource fields may remain null when they cannot be recovered without policy reconstruction.
      properties:
        participant_id:
          $ref: '#/components/schemas/ParticipantId'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        principal:
          anyOf:
            - $ref: '#/components/schemas/Principal'
            - type: 'null'
        capability:
          type: [string, 'null']
        resource:
          type: [string, 'null']
        source:
          type: string
        reason:
          type: string
        grant_id:
          type: [integer, 'null']
          minimum: 1
    IngressAuditRecord:
      type: object
      additionalProperties: false
      required: [participant_id, delivery_id, intent_id, transport, external_ref, principal]
      properties:
        participant_id:
          $ref: '#/components/schemas/ParticipantId'
        delivery_id:
          $ref: '#/components/schemas/DeliveryId'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        transport:
          type: string
        external_ref:
          type: string
        principal:
          $ref: '#/components/schemas/Principal'
    ExecutionAuditBundle:
      type: object
      additionalProperties: false
      required: [receipt, authorization, ingress]
      description: Immutable historical read model assembled from committed execution evidence.
      properties:
        receipt:
          $ref: '#/components/schemas/ExecutionReceipt'
        authorization:
          anyOf:
            - $ref: '#/components/schemas/AuthorizationProvenance'
            - type: 'null'
        ingress:
          type: array
          items:
            $ref: '#/components/schemas/IngressAuditRecord'
    ExecutionAuditEnvelope:
      type: object
      additionalProperties: false
      required: [audit]
      properties:
        audit:
          $ref: '#/components/schemas/ExecutionAuditBundle'
    ExecutionAuditIntegrityReport:
      type: object
      additionalProperties: false
      required: [participant_id, intent_id, valid, checks, violations]
      description: Read-only structural integrity report over committed execution evidence. Historical authorization is not re-evaluated.
      properties:
        participant_id:
          $ref: '#/components/schemas/ParticipantId'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        valid:
          type: boolean
        checks:
          type: array
          items:
            type: string
        violations:
          type: array
          items:
            type: string
    ExecutionAuditIntegrityEnvelope:
      type: object
      additionalProperties: false
      required: [integrity]
      properties:
        integrity:
          $ref: '#/components/schemas/ExecutionAuditIntegrityReport'
    ExecutionAuditOrphanEvidence:
      type: object
      additionalProperties: false
      required: [kind, participant_id, intent_id, reference]
      description: Read-only classification of durable execution evidence that cannot be attached to a committed receipt identity.
      properties:
        kind:
          type: string
        participant_id:
          anyOf:
            - $ref: '#/components/schemas/ParticipantId'
            - type: 'null'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        reference:
          type: [string, 'null']
    ExecutionAuditSweepReport:
      type: object
      additionalProperties: false
      required: [valid, executions_scanned, invalid_executions, orphan_evidence]
      description: Canonical database-wide read-only execution audit integrity sweep.
      properties:
        valid:
          type: boolean
        executions_scanned:
          type: integer
          minimum: 0
        invalid_executions:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionAuditIntegrityReport'
        orphan_evidence:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionAuditOrphanEvidence'
    ExecutionAuditSweepEnvelope:
      type: object
      additionalProperties: false
      required: [sweep]
      properties:
        sweep:
          $ref: '#/components/schemas/ExecutionAuditSweepReport'
    Identity:
      type: object
      required: [source, instance]
      properties:
        source:
          type: string
        instance:
          type: string
        label:
          type: [string, 'null']
        role:
          type: string
        session_type:
          type: string
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string