# src/access_api.rs
use axum::{
    extract::{Path, State},
    http::{header, HeaderMap, HeaderValue, StatusCode, Uri},
    response::{IntoResponse, Response},
    routing::get,
    Json, Router,
};
use serde_json::json;

use crate::{
    authorization, execution, http::AppState, identity, model::Identity, request_auth, web_auth,
};

pub fn app(state: AppState) -> Router {
    Router::new()
        .route("/api/access-context", get(access_context))
        .route("/api/executions/{intent_id}", get(execution_receipt))
        .route("/api/executions/{intent_id}/audit", get(execution_audit))
        .route(
            "/api/executions/{intent_id}/audit/integrity",
            get(execution_audit_integrity),
        )
        .route("/api/execution-audit/sweep", get(execution_audit_sweep))
        .with_state(state)
}

async fn access_context(
    State(state): State<AppState>,
    headers: HeaderMap,
    uri: Uri,
) -> Result<Response, AccessApiError> {
    let resolved = require_identity_for_target(&state, &headers, "GET", &uri).await?;
    let principal = principal_for_headers(&headers, &resolved);
    let lookup = resolved.instance.clone();
    let principal_for_grants = principal.clone();
    let (role, grants) = with_db(&state, move |conn| {
        let role = identity::get_web_participant_role(conn, &lookup)?;
        let grants = if role.is_some() {
            authorization::effective_grants(conn, &principal_for_grants, &lookup)?
        } else {
            Vec::new()
        };
        Ok((role, grants))
    })
    .await?;

    let participant_id = role.as_ref().map(|_| resolved.instance.clone());
    let mut capabilities = grants
        .iter()
        .map(|grant| grant.capability.clone())
        .collect::<Vec<_>>();
    capabilities.sort();
    capabilities.dedup();

    Ok(json_response(
        StatusCode::OK,
        json!({
            "identity": {
                "source": resolved.source,
                "instance": resolved.instance,
                "label": resolved.label,
            },
            "principal": principal,
            "participant_id": participant_id,
            "role": role.unwrap_or_else(|| "user".to_owned()),
            "capabilities": capabilities,
            "grants": grants,
        }),
    ))
}

async fn execution_receipt(
    State(state): State<AppState>,
    headers: HeaderMap,
    uri: Uri,
    Path(intent_id): Path<String>,
) -> Result<Response, AccessApiError> {
    let resolved = require_identity_for_target(&state, &headers, "GET", &uri).await?;
    let intent_id = execution::normalize_intent_id(&intent_id)
        .map_err(|_| AccessApiError::new(StatusCode::BAD_REQUEST, "invalid_intent_id"))?;
    let participant_id = resolved.instance.clone();
    let receipt = with_db(&state, move |conn| {
        execution::get_execution_receipt(conn, &participant_id, &intent_id)
    })
    .await?
    .ok_or_else(|| AccessApiError::new(StatusCode::NOT_FOUND, "execution_not_found"))?;

    Ok(json_response(StatusCode::OK, json!({"execution": receipt})))
}

async fn execution_audit(
    State(state): State<AppState>,
    headers: HeaderMap,
    uri: Uri,
    Path(intent_id): Path<String>,
) -> Result<Response, AccessApiError> {
    let resolved = require_identity_for_target(&state, &headers, "GET", &uri).await?;
    let principal = principal_for_headers(&headers, &resolved);
    let intent_id = execution::normalize_intent_id(&intent_id)
        .map_err(|_| AccessApiError::new(StatusCode::BAD_REQUEST, "invalid_intent_id"))?;
    let participant_id = resolved.instance.clone();
    let policy_principal = principal.clone();
    let lookup_participant = participant_id.clone();
    let lookup_intent = intent_id.clone();
    let (allowed, audit) = with_db(&state, move |conn| {
        let allowed = authorization::authorize(
            conn,
            &policy_principal,
            &lookup_participant,
            authorization::READ_EXECUTION_AUDIT,
            Some(&lookup_intent),
        )?;
        let audit = if allowed {
            execution::get_execution_audit_bundle(conn, &lookup_participant, &lookup_intent)?
        } else {
            None
        };
        Ok((allowed, audit))
    })
    .await?;
    if !allowed {
        return Err(AccessApiError::new(StatusCode::FORBIDDEN, "forbidden"));
    }
    let audit =
        audit.ok_or_else(|| AccessApiError::new(StatusCode::NOT_FOUND, "execution_not_found"))?;
    Ok(json_response(StatusCode::OK, json!({"audit": audit})))
}

async fn execution_audit_integrity(
    State(state): State<AppState>,
    headers: HeaderMap,
    uri: Uri,
    Path(intent_id): Path<String>,
) -> Result<Response, AccessApiError> {
    let resolved = require_identity_for_target(&state, &headers, "GET", &uri).await?;
    let principal = principal_for_headers(&headers, &resolved);
    let intent_id = execution::normalize_intent_id(&intent_id)
        .map_err(|_| AccessApiError::new(StatusCode::BAD_REQUEST, "invalid_intent_id"))?;
    let participant_id = resolved.instance.clone();
    let policy_principal = principal.clone();
    let lookup_participant = participant_id.clone();
    let lookup_intent = intent_id.clone();
    let (allowed, report) = with_db(&state, move |conn| {
        let allowed = authorization::authorize(
            conn,
            &policy_principal,
            &lookup_participant,
            authorization::READ_EXECUTION_AUDIT,
            Some(&lookup_intent),
        )?;
        let report = if allowed {
            Some(execution::verify_execution_audit_integrity(
                conn,
                &lookup_participant,
                &lookup_intent,
            )?)
        } else {
            None
        };
        Ok((allowed, report))
    })
    .await?;
    if !allowed {
        return Err(AccessApiError::new(StatusCode::FORBIDDEN, "forbidden"));
    }
    let report = report.expect("authorized integrity verification must produce a report");
    Ok(json_response(StatusCode::OK, json!({"integrity": report})))
}

async fn execution_audit_sweep(
    State(state): State<AppState>,
    headers: HeaderMap,
    uri: Uri,
) -> Result<Response, AccessApiError> {
    let resolved = require_identity_for_target(&state, &headers, "GET", &uri).await?;
    let principal = principal_for_headers(&headers, &resolved);
    let participant_id = resolved.instance.clone();
    let policy_principal = principal.clone();
    let lookup_participant = participant_id.clone();
    let (allowed, report) = with_db(&state, move |conn| {
        let allowed = authorization::authorize(
            conn,
            &policy_principal,
            &lookup_participant,
            authorization::READ_EXECUTION_AUDIT_SWEEP,
            Some(authorization::EXECUTION_AUDIT_SWEEP_RESOURCE),
        )?;
        let report = if allowed {
            Some(execution::sweep_execution_audit_integrity(conn)?)
        } else {
            None
        };
        Ok((allowed, report))
    })
    .await?;
    if !allowed {
        return Err(AccessApiError::new(StatusCode::FORBIDDEN, "forbidden"));
    }
    let report = report.expect("authorized audit sweep must produce a report");
    Ok(json_response(StatusCode::OK, json!({"sweep": report})))
}

fn principal_for_headers(headers: &HeaderMap, resolved: &Identity) -> execution::Principal {
    let provider = if headers.contains_key(web_auth::WEB_SESSION_HEADER) {
        "human-web"
    } else if headers.contains_key(request_auth::AUTH_PROOF_HEADER)
        || headers.contains_key(request_auth::AUTH_SCHEME_HEADER)
    {
        "participant-hmac"
    } else {
        "bearer"
    };
    execution::Principal {
        provider: provider.to_owned(),
        subject: resolved.instance.clone(),
    }
}

async fn require_identity_for_target(
    state: &AppState,
    headers: &HeaderMap,
    method: &'static str,
    uri: &Uri,
) -> Result<Identity, AccessApiError> {
    let headers = headers.clone();
    let request_target = uri
        .path_and_query()
        .map(|value| value.as_str())
        .unwrap_or_else(|| uri.path())
        .to_owned();
    with_db(state, move |conn| {
        request_auth::resolve_request_identity_for_target(conn, &headers, method, &request_target)
    })
    .await?
    .ok_or_else(AccessApiError::unauthorized)
}

async fn with_db<T, F>(state: &AppState, operation: F) -> Result<T, AccessApiError>
where
    T: Send + 'static,
    F: FnOnce(&rusqlite::Connection) -> rusqlite::Result<T> + Send + 'static,
{
    let path = state.db_path.clone();
    tokio::task::spawn_blocking(move || {
        let conn = crate::db::connect(&path)?;
        operation(&conn)
    })
    .await
    .map_err(|_| AccessApiError::new(StatusCode::INTERNAL_SERVER_ERROR, "internal_error"))?
    .map_err(|_| AccessApiError::new(StatusCode::SERVICE_UNAVAILABLE, "database_unavailable"))
}

#[derive(Debug)]
struct AccessApiError {
    status: StatusCode,
    code: &'static str,
}

impl AccessApiError {
    fn new(status: StatusCode, code: &'static str) -> Self {
        Self { status, code }
    }

    fn unauthorized() -> Self {
        Self::new(StatusCode::UNAUTHORIZED, "unauthorized")
    }
}

impl IntoResponse for AccessApiError {
    fn into_response(self) -> Response {
        json_response(self.status, json!({"error": self.code}))
    }
}

fn json_response(status: StatusCode, value: serde_json::Value) -> Response {
    let mut response = (status, Json(value)).into_response();
    response
        .headers_mut()
        .insert(header::CACHE_CONTROL, HeaderValue::from_static("no-store"));
    response
}